Creating an Algorithm for the game
In this function, I study an algorithm that can define the probability of having Covid-19.
Probability is given by moltiplication of many moltiplicators.

The point is to give an idea on the probability, in future I will make the moltiplicators more accurates, at the moment are just some intuitives numbers.